pyproject.toml
[project]
name = "e2e-flow"
version = "0.1.0"
description = "E2E Flow with Prefect worker on Snowflake SPCS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "prefect>=3.6.4",
    "prefect-snowflake>=0.28.7",
    "psycopg[binary]>=3.3.1",
    "python-dotenv>=1.2.1",
    "pandas>=2.3.3",
    "pandera>=0.27.0",
]
